Declare some metods as statict
The class ConnectorBase has some private methods that do not use the class members or functions.
All these methods should be declared as static or moved in .cpp file,
because they do not use/change any class members.
Plus, there is a method getEDID of class ConnectorBase.
The method returns the block size and does not change the class state, the 'const is added
to the method declaration.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>